Refactor report config downloads to make way for DPS downloads as well
