+++ server/routes/downloadReportConfig/nomis.ts
@@ -1,59 +1,12 @@
 import { Router, type Request, type Response } from 'express'
 import { NotFound } from 'http-errors'
 
-import type { ReferenceCode } from '../data/prisonApi'
-import format from '../utils/format'
+import type { ReferenceCode } from '../../data/prisonApi'
+import format from '../../utils/format'
+import { jsonDownload, streamCsvDownload, type CsvCellValue } from './utils'
 
 function isFileTypeSupported(fileType: string): fileType is 'csv' | 'json' {
   return ['csv', 'json'].includes(fileType)
-}
-
-function jsonDownload<T>(res: Response, fileName: string, data: T): void {
-  res.header('Content-Type', 'application/json')
-  res.header('Content-Disposition', `attachment; filename="${fileName}.json"`)
-  res.send(JSON.stringify(data))
-  res.end()
-}
-
-type CsvCellValue = string | number | boolean | null
-
-function streamCsvDownload(res: Response, fileName: string, lines: Generator<CsvCellValue[]>): void {
-  res.header('Content-Type', 'text/csv')
-  res.header('Content-Disposition', `attachment; filename="${fileName}.csv"`)
-  for (const line of lines) {
-    const lineString = line.map(escapeCsvCell).join(',')
-    res.write(`${lineString}\n`)
-  }
-  res.end()
-}
-
-function escapeCsvCell(value: CsvCellValue): string {
-  /* eslint-disable no-param-reassign */
-  if (value === null || value === undefined) {
-    return ''
-  }
-  if (value === true) {
-    return 'TRUE'
-  }
-  if (value === false) {
-    return 'FALSE'
-  }
-  let mustQuote = false
-  if (typeof value === 'number') {
-    value = value.toString()
-  }
-  if (value.includes(',')) {
-    mustQuote = true
-  }
-  if (value.includes('"')) {
-    value = value.replaceAll('"', '""')
-    mustQuote = true
-  }
-  if (mustQuote) {
-    return `"${value}"`
-  }
-  return value
-  /* eslint-enable no-param-reassign */
 }
 
 async function downloadIncidentTypes(req: Request, res: Response): Promise<void> {
